Turn resolvers/Mutation.ts file into directory
